+++ controllers/contacts/patchFavorite.js
@@ -6,14 +6,14 @@ const patchFavorite = async (req, res, next) => {
         const { contactId } = req.params;
         const { favorite } = req.body;
 
+        if (!req.body) {
+            throw createError(400, "Missing field favorite");
+        }
+
         const result = await Contact.findByIdAndUpdate(contactId, { favorite }, { new: true });
 
         if (!result) {
             throw createError(404, `UPS! Contact with id=${contactId} not found`);
-        }
-
-        if (!req.body) {
-            throw createError(400, "Missing field favorite");
         }
 
         res.json({
