+++ model/contact.js
@@ -1,5 +1,4 @@
 const { Schema, model } = require("mongoose");
-const Joi = require('joi');
 
 const contactSchema = Schema({
     name: {
@@ -16,24 +15,10 @@ const contactSchema = Schema({
         type: Boolean,
         default: false,
     },
-
 }, { versionKey: false, timeStamps: true });
-
-const joiSchema = Joi.object({
-    name: Joi.string().min(3).required(),
-    email: Joi.string().email().required(),
-    phone: Joi.string().min(7).required(),
-});
-    
-const favoriteJoiSchema = Joi.object({
-    favorite: Joi.bool().required(),
-});
-
 
 const Contact = model("contact", contactSchema);
 
 module.exports = {
     Contact,
-    joiSchema,
-    favoriteJoiSchema
 }
