+++ model/index.js
@@ -1,9 +1,7 @@
 const {Contact} = require("./contact");
-const joiSchema = require("./contact");
-const favoriteJoiSchema = require("./contact");
+
 
 module.exports = {
     Contact,
-    joiSchema, 
-    favoriteJoiSchema
+    
 };
