+++ validation/contacts/contactsValidate.js
@@ -1,26 +1,5 @@
 const Joi = require('joi');
 
-// const validation =  (req, res, next) => {
-    
-//     const contactSchema = Joi.object({
-//     name: Joi.string().min(3).required(),
-//     email: Joi.string().email().required(),
-//     phone: Joi.string().min(7).required(),
-//     });
-    
-//     const validationResult = contactSchema.validate(req.body);
-
-//     if (validationResult.error) {
-//         return res.status(400)
-//             .json({
-//                 'message': validationResult.error.message,
-//                 'status': 400,
-//             });
-//     }
-
-//     next();
-// };
-
 const validation = (schema) => {
     return (req, res, next) => {
         const { error } = schema.validate(req.body);
@@ -31,6 +10,5 @@ const validation = (schema) => {
         next();
     };
 };
-
 
 module.exports = validation;
